Revert "Feature: Adds connectWithCleanup HOC (#19392)" (#19628)
This reverts commit 989f98efdadb2000423dc672effdc65e6de944e8.
