+++ public/app/types/store.ts
@@ -16,7 +16,6 @@ import { NavIndex } from '@grafana/data';
 import { ApplicationState } from './application';
 import { LdapState, LdapUserState } from './ldap';
 import { PanelEditorState } from '../features/dashboard/panel_editor/state/reducers';
-import { ApiKeysState } from './apiKeys';
 
 export interface StoreState {
   navIndex: NavIndex;
@@ -37,7 +36,6 @@ export interface StoreState {
   application: ApplicationState;
   ldap: LdapState;
   ldapUser: LdapUserState;
-  apiKeys: ApiKeysState;
 }
 
 /*
